refactor: move modals, onboarding, projects landing, notifications into modules
