+++ src/render.rs
@@ -405,309 +405,8 @@ impl Render for Kyde {
 }
 
 impl Kyde {
-    fn render_projects(
-        &self,
-        ui: &'static str,
-        fs: gpui::Pixels,
-        cx: &mut Context<Self>,
-    ) -> gpui::AnyElement {
-        // No recent projects → the animated welcome hero (no search box, nothing to search).
-        if self.recents.paths.is_empty() {
-            return self.render_welcome(ui, fs, cx);
-        }
-        let query = self.project_search.read(cx).text().to_lowercase();
 
-        // Primary = filled accent + white text. Secondary = transparent bg, divider
-        // border, secondary text.
-        let button = |label: &'static str, accent: bool, _cx: &mut Context<Self>| {
-            let t = theme::get();
-            div()
-                .px_4()
-                .py_1p5()
-                .rounded_md()
-                .border_1()
-                .font_weight(FontWeight::SEMIBOLD)
-                .when(accent, |d| d.bg(t.primary).text_color(t.primary_text))
-                .when(!accent, |d| {
-                    d.border_color(t.divider).text_color(t.secondary_text)
-                })
-                .child(label)
-        };
 
-        // A draggable title strip holds the macOS traffic lights on its own line, so the search
-        // row sits cleanly *below* them (not tucked under the close/minimise buttons).
-        let titlebar = div().flex().flex_none().h(px(38.0)).w_full().child(
-            div()
-                .size_full()
-                .window_control_area(gpui::WindowControlArea::Drag),
-        );
-        // One row below the titlebar: search box fills the left, New Project / Open inline right.
-        let header = div()
-            .flex()
-            .flex_row()
-            .items_center()
-            .gap_2()
-            .px_4()
-            .py_2()
-            .border_b_1()
-            .border_color(theme::get().divider)
-            .child(
-                div()
-                    .flex_1()
-                    .flex()
-                    .flex_row()
-                    .items_center()
-                    .gap_2()
-                    .text_color(theme::get().line_number)
-                    .child(
-                        svg()
-                            .path("icons/search.svg")
-                            .size(px(15.0))
-                            .text_color(theme::get().line_number),
-                    )
-                    .child(div().flex_1().child(self.project_search.clone())),
-            )
-            .child(button("New Project", false, cx).on_mouse_down(
-                MouseButton::Left,
-                cx.listener(|this, _e, _w, cx| this.pick_folder(cx)),
-            ))
-            .child(button("Open", true, cx).on_mouse_down(
-                MouseButton::Left,
-                cx.listener(|this, _e, _w, cx| this.pick_folder(cx)),
-            ));
-
-        let body = self.render_recents_list(&query, cx);
-
-        let mut root = div()
-            .key_context("Kyde")
-            .track_focus(&self.focus_handle)
-            .relative()
-            .flex()
-            .flex_col()
-            .size_full()
-            .bg(theme::get().main_bg)
-            .font_family(ui)
-            .text_size(fs)
-            .child(titlebar)
-            .child(header)
-            .child(body)
-            // Git-op error + crash banners pinned to the bottom of the window.
-            .when(self.op_error.is_some(), |d| {
-                d.child(self.render_op_error_banner(ui, cx))
-            })
-            .when(self.pending_crash.is_some(), |d| {
-                d.child(self.render_crash_banner(ui, cx))
-            });
-
-        if self.onboarding_open {
-            root = root.child(self.render_onboarding(ui, fs, cx));
-        }
-        root.into_any_element()
-    }
-
-    /// The scrollable recent-projects list (Projects screen, when there are recents).
-    fn render_recents_list(&self, query: &str, cx: &mut Context<Self>) -> gpui::AnyElement {
-        let rows = self
-            .recents
-            .paths
-            .iter()
-            .filter(|p| {
-                if query.is_empty() {
-                    return true;
-                }
-                let name = projects::name_of(p).to_lowercase();
-                name.contains(query) || p.to_string_lossy().to_lowercase().contains(query)
-            })
-            .map(|p| {
-                let name = projects::name_of(p);
-                let icon_color = gpui::rgb(projects::color_for(&name));
-                let pretty = projects::pretty_path(p);
-                let pclone = p.clone();
-                div()
-                    .flex()
-                    .flex_row()
-                    .items_center()
-                    .gap_3()
-                    .px_4()
-                    .py_2()
-                    .hover(|s| s.bg(theme::get().caret_row))
-                    .child(
-                        div()
-                            .flex()
-                            .items_center()
-                            .justify_center()
-                            .size(px(34.0))
-                            .rounded_md()
-                            .bg(icon_color)
-                            .text_color(gpui::white())
-                            .child(SharedString::from(projects::initials(&name))),
-                    )
-                    .child(
-                        div()
-                            .flex()
-                            .flex_col()
-                            .child(
-                                div()
-                                    .font_weight(FontWeight::SEMIBOLD)
-                                    .text_color(theme::get().text)
-                                    .child(SharedString::from(name)),
-                            )
-                            .child(
-                                div()
-                                    .text_color(theme::get().line_number)
-                                    .child(SharedString::from(pretty)),
-                            ),
-                    )
-                    .on_mouse_down(
-                        MouseButton::Left,
-                        cx.listener(move |this, _e, _w, cx| this.open_project(pclone.clone(), cx)),
-                    )
-            });
-        div()
-            .id("recents")
-            .overflow_y_scroll()
-            .flex()
-            .flex_col()
-            .flex_1()
-            .children(rows)
-            .into_any_element()
-    }
-
-    /// First-run / no-recents welcome: an animated 3D "KY" (ANSI-Shadow blocks with a
-    /// diagonal shimmer sweeping the faces), a tagline, and New Project / Open Folder.
-    fn render_welcome(
-        &self,
-        ui: &'static str,
-        fs: gpui::Pixels,
-        cx: &mut Context<Self>,
-    ) -> gpui::AnyElement {
-        let t = theme::get();
-        const ART: &[&str] = &[
-            "██╗  ██╗██╗   ██╗██████╗ ███████╗",
-            "██║ ██╔╝╚██╗ ██╔╝██╔══██╗██╔════╝",
-            "█████╔╝  ╚████╔╝ ██║  ██║█████╗  ",
-            "██╔═██╗   ╚██╔╝  ██║  ██║██╔══╝  ",
-            "██║  ██╗   ██║   ██████╔╝███████╗",
-            "╚═╝  ╚═╝   ╚═╝   ╚═════╝ ╚══════╝",
-        ];
-        let frame = self.welcome_frame as f32;
-        let mono = gpui::Font {
-            family: theme::font::FAMILY.into(),
-            features: Default::default(),
-            fallbacks: None,
-            weight: FontWeight::BOLD,
-            style: Default::default(),
-        };
-        let shadow: gpui::Hsla = gpui::rgb(0x223056).into();
-        let art_lines: Vec<gpui::AnyElement> = ART
-            .iter()
-            .enumerate()
-            .map(|(row, line)| {
-                let runs: Vec<gpui::TextRun> = line
-                    .chars()
-                    .enumerate()
-                    .map(|(col, ch)| {
-                        let color: gpui::Hsla = if ch == '█' {
-                            // diagonal highlight band sweeping left→right over the faces
-                            let phase = (col as f32 * 0.5 + row as f32 * 0.45) - frame * 0.14;
-                            let b = phase.sin() * 0.5 + 0.5;
-                            lerp_rgb(0x2E5BD0, 0xCFE0FF, b).into()
-                        } else if ch == ' ' {
-                            gpui::rgba(0x00000000).into()
-                        } else {
-                            shadow
-                        };
-                        gpui::TextRun {
-                            len: ch.len_utf8(),
-                            font: mono.clone(),
-                            color,
-                            background_color: None,
-                            underline: None,
-                            strikethrough: None,
-                        }
-                    })
-                    .collect();
-                gpui::StyledText::new(SharedString::from(*line))
-                    .with_runs(runs)
-                    .into_any_element()
-            })
-            .collect();
-
-        let new_btn = btn_primary("welcome-new", "New Project")
-            .w_full()
-            .flex()
-            .justify_center()
-            .on_mouse_down(
-                MouseButton::Left,
-                cx.listener(|this, _e, _w, cx| this.pick_folder(cx)),
-            );
-        let open_btn = btn_secondary("welcome-open", "Open Folder")
-            .w_full()
-            .flex()
-            .justify_center()
-            .on_mouse_down(
-                MouseButton::Left,
-                cx.listener(|this, _e, _w, cx| this.pick_folder(cx)),
-            );
-
-        let hero = div()
-            .flex_1()
-            .flex()
-            .flex_col()
-            .items_center()
-            .justify_center()
-            .gap_6()
-            .child(
-                div()
-                    .flex()
-                    .flex_col()
-                    .items_center()
-                    .text_size(px(16.0))
-                    .line_height(px(18.0))
-                    .children(art_lines),
-            )
-            .child(
-                div()
-                    .flex()
-                    .flex_col()
-                    .gap_2()
-                    .w(px(220.0))
-                    .mt_4()
-                    .child(new_btn)
-                    .child(open_btn),
-            );
-
-        let mut root = div()
-            .key_context("Kyde")
-            .track_focus(&self.focus_handle)
-            .relative()
-            .flex()
-            .flex_col()
-            .size_full()
-            .bg(t.frame_bg)
-            .font_family(ui)
-            .text_size(fs)
-            // Draggable strip over the traffic lights (the window has a transparent titlebar).
-            .child(
-                div()
-                    .h(px(40.0))
-                    .flex_none()
-                    .w_full()
-                    .window_control_area(gpui::WindowControlArea::Drag),
-            )
-            .child(hero)
-            // Git-op error + crash banners pinned to the bottom of the window.
-            .when(self.op_error.is_some(), |d| {
-                d.child(self.render_op_error_banner(ui, cx))
-            })
-            .when(self.pending_crash.is_some(), |d| {
-                d.child(self.render_crash_banner(ui, cx))
-            });
-        if self.onboarding_open {
-            root = root.child(self.render_onboarding(ui, fs, cx));
-        }
-        root.into_any_element()
-    }
 
 
 
@@ -1722,92 +1421,6 @@ impl Kyde {
     }
 
 
-    /// Editor tab strip: one tab per open file, left→right in open order. Click activates,
-    /// the `×` closes, right-click opens the tab context menu (close / others / to the right).
-    /// Project tabs strip: one pill per open project, above all other UI (under the title
-    /// bar). Click switches projects; the `×` closes one. Rendered only when >1 is open.
-    fn render_project_tabs(&self, ui: &'static str, cx: &mut Context<Self>) -> gpui::AnyElement {
-        let t = theme::get();
-        let tabs = self.open_projects.iter().enumerate().map(|(i, root)| {
-            let active = self.repo_root.as_ref() == Some(root);
-            let name: SharedString = crate::projects::name_of(root).into();
-            let grp = SharedString::from(format!("projgrp-{i}"));
-            let close = div()
-                .id(SharedString::from(format!("project-close-{i}")))
-                .flex_none()
-                .w(px(18.0))
-                .flex()
-                .items_center()
-                .justify_center()
-                .rounded_sm()
-                .text_size(px(15.0))
-                .text_color(t.line_number)
-                .hover(|d| d.bg(t.bg_light).text_color(t.text))
-                .when(!active, |d| {
-                    d.opacity(0.0).group_hover(grp.clone(), |s| s.opacity(1.0))
-                })
-                .child("×")
-                .on_mouse_down(
-                    MouseButton::Left,
-                    cx.listener({
-                        let root = root.clone();
-                        move |this, _e, _w, cx| {
-                            cx.stop_propagation();
-                            this.close_project(root.clone(), cx);
-                        }
-                    }),
-                );
-            div()
-                .id(SharedString::from(format!("project-tab-{i}")))
-                .group(grp.clone())
-                .flex()
-                .flex_row()
-                .items_center()
-                .gap_2()
-                .px_3()
-                .h(px(28.0))
-                .flex_none()
-                .max_w(px(220.0))
-                .rounded_md()
-                .border_1()
-                .cursor_pointer()
-                .when(active, |d| {
-                    d.bg(gpui::rgba(0x3574F026)).border_color(t.primary)
-                })
-                .when(!active, |d| {
-                    d.border_color(gpui::rgba(0x00000000))
-                        .hover(|d| d.bg(t.bg_mid))
-                })
-                .text_color(if active { t.text } else { t.line_number })
-                .child(div().min_w_0().truncate().child(name))
-                .child(close)
-                .on_mouse_down(
-                    MouseButton::Left,
-                    cx.listener({
-                        let root = root.clone();
-                        move |this, _e, _w, cx| {
-                            if this.repo_root.as_ref() != Some(&root) {
-                                this.open_project(root.clone(), cx);
-                            }
-                        }
-                    }),
-                )
-        });
-        div()
-            .flex()
-            .flex_row()
-            .items_center()
-            .gap_1()
-            .flex_none()
-            .w_full()
-            .px_2()
-            .pb(px(theme::FRAME_GAP))
-            .bg(t.frame_bg)
-            .font_family(ui)
-            .text_size(px(t.ui_font_size))
-            .children(tabs)
-            .into_any_element()
-    }
 
     fn render_tab_bar(
         &self,
@@ -2084,496 +1697,11 @@ impl Kyde {
             .into_any_element()
     }
 
-    /// Top-of-editor prompt offering to install syntax support for the open file.
-    /// IntelliJ-style: a thin bar with a primary (#3473EE) Install button.
-    /// Top-of-window banner shown only when a newer release exists. The action is
-    /// "Update & Relaunch" when running from a `.app` bundle (downloads + swaps in place),
-    /// else "Download" (opens the release page) — see `do_update`.
-    fn render_update_banner(&self, ui: &'static str, cx: &mut Context<Self>) -> gpui::AnyElement {
-        let t = theme::get();
-        let Some(rel) = self.update_available.as_ref() else {
-            return div().into_any_element();
-        };
-        let can_swap = update::running_bundle().is_some() && !rel.zip_url.is_empty();
-        let action_label = if self.updating {
-            "Updating…"
-        } else if can_swap {
-            "Update & Relaunch"
-        } else {
-            "Download"
-        };
-        let msg: SharedString = format!("Update available — v{}", rel.version).into();
-
-        // ↑ badge
-        let badge = div()
-            .flex_none()
-            .size(px(18.0))
-            .flex()
-            .items_center()
-            .justify_center()
-            .rounded_full()
-            .bg(t.primary)
-            .child(
-                div()
-                    .text_size(px(12.0))
-                    .font_weight(FontWeight::BOLD)
-                    .text_color(gpui::white())
-                    .child("↑"),
-            );
-
-        let updating = self.updating;
-        div()
-            .flex()
-            .flex_row()
-            .items_center()
-            .gap_3()
-            .px_3()
-            .py_2()
-            .bg(t.bg_mid)
-            .border_b_1()
-            .border_color(t.divider)
-            .font_family(ui)
-            .text_size(px(t.ui_font_size))
-            .text_color(t.text)
-            .child(
-                div()
-                    .flex()
-                    .flex_row()
-                    .items_center()
-                    .gap_2()
-                    .child(badge)
-                    .child(msg),
-            )
-            // Spacer pushes the actions to the right.
-            .child(div().flex_1().min_w_0())
-            .child(
-                btn_primary("update-now", action_label)
-                    .when(updating, |d| d.opacity(0.6))
-                    .on_mouse_down(
-                        MouseButton::Left,
-                        cx.listener(|this, _e, _w, cx| this.do_update(cx)),
-                    ),
-            )
-            .child(btn_secondary("update-dismiss", "Dismiss").on_mouse_down(
-                MouseButton::Left,
-                cx.listener(|this, _e, _w, cx| this.dismiss_update(cx)),
-            ))
-            .into_any_element()
-    }
-
-    fn render_install_banner(
-        &self,
-        pack: &'static highlight::Pack,
-        ui: &'static str,
-        fs: gpui::Pixels,
-        cx: &mut Context<Self>,
-    ) -> gpui::AnyElement {
-        let t = theme::get();
-        let ext = self
-            .open_path
-            .as_ref()
-            .and_then(|p| p.extension())
-            .and_then(|e| e.to_str())
-            .unwrap_or("");
-
-        // ⓘ info badge
-        let info = div()
-            .flex_none()
-            .size(px(18.0))
-            .flex()
-            .items_center()
-            .justify_center()
-            .rounded_full()
-            .bg(t.primary)
-            .child(
-                div()
-                    .text_size(px(11.0))
-                    .font_weight(FontWeight::BOLD)
-                    .text_color(gpui::white())
-                    .child("i"),
-            );
-
-        let link = |label: SharedString, id: &'static str| {
-            div()
-                .id(id)
-                .flex_none()
-                .text_color(t.primary)
-                .cursor_pointer()
-                .hover(|s| s.text_color(t.text))
-                .child(label)
-        };
-
-        div()
-            .flex()
-            .flex_row()
-            .items_center()
-            .gap_4()
-            .px_3()
-            .py_2()
-            .bg(t.bg_mid)
-            .border_b_1()
-            .border_color(t.divider)
-            .font_family(ui)
-            .text_size(fs)
-            .text_color(t.text)
-            .child(
-                div()
-                    .flex()
-                    .flex_row()
-                    .items_center()
-                    .gap_2()
-                    .child(info)
-                    .child(SharedString::from(format!("Plugins supporting *.{ext}"))),
-            )
-            .child(
-                link(
-                    format!("Install {} plugin", pack.name).into(),
-                    "install-plugin",
-                )
-                .on_mouse_down(
-                    MouseButton::Left,
-                    cx.listener(|this, _e, _w, cx| {
-                        this.install_open_pack(cx);
-                        cx.notify();
-                    }),
-                ),
-            )
-            .child(link("Ignore extension".into(), "ignore-ext").on_mouse_down(
-                MouseButton::Left,
-                cx.listener(|this, _e, _w, cx| this.ignore_open_pack(cx)),
-            ))
-            .into_any_element()
-    }
 
 
-    /// The language-plugin manager: every installable pack with its monogram badge,
-    /// approximate compiled footprint, install state, and an Install/Uninstall button.
-    /// Filtered live by the search box.
-    /// Body of the Language-Plugins modal window (hosted by `ModalWindow`, native titlebar).
-    pub(crate) fn render_plugins_body(&mut self, cx: &mut Context<Self>) -> gpui::AnyElement {
-        let ui = theme::font::UI_FAMILY;
-        let t = theme::get();
-        let q = self.plugins_query.read(cx).text().to_lowercase();
-        let q = q.trim();
-        let mut packs: Vec<_> = highlight::PACKS
-            .iter()
-            .filter(|p| q.is_empty() || p.name.to_lowercase().contains(q) || p.id.contains(q))
-            .collect();
-        packs.sort_by_key(|p| p.name.to_lowercase());
-        let rows: Vec<gpui::AnyElement> = packs
-            .into_iter()
-            .map(|p| {
-                let installed = self.plugins.is_installed(p.id);
-                let id = p.id;
-                // Larger badge: the rows are two lines tall (name + size), so size the
-                // language monogram to roughly match.
-                let badge = badge_inner(
-                    file_badge(std::path::Path::new(&format!("x.{}", pack_ext(p.id)))),
-                    14.0,
-                );
-                let bid = SharedString::from(format!("pkg-{id}"));
-                let btn = if installed {
-                    btn_secondary(bid, "Uninstall")
-                } else {
-                    btn_primary(bid, "Install")
-                }
-                .flex_none()
-                .on_mouse_down(
-                    MouseButton::Left,
-                    cx.listener(move |this, _e, _w, cx| this.toggle_plugin(id, cx)),
-                );
-                div()
-                    .flex()
-                    .flex_row()
-                    .items_center()
-                    .gap_3()
-                    .px_3()
-                    .py_2()
-                    .rounded_md()
-                    .hover(|s| s.bg(t.bg_mid))
-                    .child(
-                        div()
-                            .w(px(40.0))
-                            .flex_none()
-                            .flex()
-                            .items_center()
-                            .justify_center()
-                            .child(badge),
-                    )
-                    .child(
-                        div()
-                            .flex_1()
-                            .min_w_0()
-                            .flex()
-                            .flex_col()
-                            .child(div().text_color(t.text).child(SharedString::from(p.name)))
-                            .child(
-                                div()
-                                    .text_color(t.line_number)
-                                    .text_size(px(11.0))
-                                    .child(SharedString::from(pack_size(p.id))),
-                            ),
-                    )
-                    .child(btn)
-                    .into_any_element()
-            })
-            .collect();
 
-        // Fills the modal window (chrome + native titlebar come from `ModalWindow`).
-        div()
-            .size_full()
-            .flex()
-            .flex_col()
-            .font_family(ui)
-            .child(div().px_2().py_2().child(self.plugins_query.clone()))
-            // Divider under the search input.
-            .child(div().h(px(1.0)).bg(t.divider))
-            .child(
-                div()
-                    .id("plugins-list")
-                    .overflow_y_scroll()
-                    .flex_1()
-                    .flex()
-                    .flex_col()
-                    .p_1()
-                    .children(rows),
-            )
-            .into_any_element()
-    }
 
-    fn render_onboarding(
-        &self,
-        ui: &'static str,
-        fs: gpui::Pixels,
-        cx: &mut Context<Self>,
-    ) -> gpui::AnyElement {
-        let preset_card = |preset: Preset, cx: &mut Context<Self>| {
-            let selected = self.onboarding_choice == preset;
-            let sample: Vec<gpui::AnyElement> = keymap::ACTIONS
-                .iter()
-                .map(|a| {
-                    let key = match preset {
-                        Preset::VSCode => a.vscode,
-                        _ => a.webstorm,
-                    };
-                    div()
-                        .flex()
-                        .flex_row()
-                        .justify_between()
-                        .gap_4()
-                        .child(SharedString::from(a.label))
-                        .child(
-                            div()
-                                .px_2()
-                                .bg(theme::get().bg_mid)
-                                .rounded_md()
-                                .text_color(theme::get().line_number)
-                                .child(SharedString::from(pretty_key(key))),
-                        )
-                        .into_any_element()
-                })
-                .collect();
-            div()
-                .flex()
-                .flex_col()
-                .gap_2()
-                .w(px(300.0))
-                .p_3()
-                .rounded_lg()
-                // Selection = thick accent border + a cool same-family gradient.
-                .border_2()
-                .border_color(if selected {
-                    theme::get().primary
-                } else {
-                    theme::get().bg_light
-                })
-                .when(selected, |d| {
-                    d.bg(gpui::linear_gradient(
-                        145.0,
-                        gpui::linear_color_stop(gpui::rgb(0x232838), 0.0),
-                        gpui::linear_color_stop(gpui::rgb(0x2E3A5C), 1.0),
-                    ))
-                })
-                .when(!selected, |d| d.bg(theme::get().panel_bg))
-                .cursor_pointer()
-                .child(
-                    div()
-                        .text_color(theme::get().text)
-                        .child(SharedString::from(preset.label())),
-                )
-                .children(sample)
-                .on_mouse_down(
-                    MouseButton::Left,
-                    cx.listener(move |this, _e, _w, cx| {
-                        this.onboarding_choice = preset;
-                        cx.notify();
-                    }),
-                )
-        };
 
-        let panel = div()
-            .flex()
-            .flex_col()
-            .gap_4()
-            .p_4()
-            .bg(theme::get().bg_mid)
-            .border_1()
-            .border_color(theme::get().bg_light)
-            .rounded_md()
-            .shadow_lg()
-            .font_family(ui)
-            .text_size(fs)
-            .text_color(theme::get().text)
-            .child(
-                div()
-                    .text_color(theme::get().text)
-                    .font_weight(FontWeight::SEMIBOLD)
-                    .child("Choose your keymap"),
-            )
-            .child(
-                div()
-                    .text_color(theme::get().line_number)
-                    .child(if self.onboarding_forced {
-                        "Pick a keymap to get started. You can change it later in Kyde → Settings."
-                    } else {
-                        "Reopen any time from Kyde → Settings (⌘,)."
-                    }),
-            )
-            .child(
-                div()
-                    .flex()
-                    .flex_row()
-                    .gap_4()
-                    .child(preset_card(Preset::WebStorm, cx))
-                    .child(preset_card(Preset::VSCode, cx)),
-            )
-            .child(self.render_shell_command_row(cx))
-            // Single primary action, bottom-right: confirm the highlighted choice.
-            .child(
-                div().flex().flex_row().justify_end().mt_2().child(
-                    div()
-                        .px_5()
-                        .py_1p5()
-                        .rounded_md()
-                        .bg(theme::get().primary)
-                        .text_color(gpui::white())
-                        .font_weight(FontWeight::SEMIBOLD)
-                        .cursor_pointer()
-                        .child("Continue")
-                        .on_mouse_down(
-                            MouseButton::Left,
-                            cx.listener(|this, _e, _w, cx| {
-                                // Apply the shell-command checkbox before closing: only
-                                // when ticked and a name is actually free to claim.
-                                if this.onboarding_install_cmd
-                                    && matches!(shellcmd::state(), shellcmd::State::Available(_))
-                                {
-                                    if let Err(e) = shellcmd::install() {
-                                        this.shell_cmd_error = Some(e);
-                                    }
-                                }
-                                let choice = this.onboarding_choice;
-                                this.choose_preset(choice, cx);
-                            }),
-                        ),
-                ),
-            );
-
-        // Clicks inside the panel must not reach the backdrop.
-        let panel = panel.on_mouse_down(MouseButton::Left, |_, _, cx| cx.stop_propagation());
-
-        overlay(cx, !self.onboarding_forced)
-            .child(panel)
-            .into_any_element()
-    }
-
-    /// One row in the keymap picker: a checkbox to install a `ky`/`kyde` shell
-    /// launcher (symlink into ~/.local/bin, VSCode-style). Shown on both first
-    /// run and reopened Settings; the symlink is created when Continue confirms.
-    /// Renders nothing when we can't resolve a location (`Unavailable`).
-    fn render_shell_command_row(&self, cx: &mut Context<Self>) -> gpui::AnyElement {
-        let t = theme::get();
-        let st = shellcmd::state();
-        if matches!(st, shellcmd::State::Unavailable) {
-            return div().into_any_element();
-        }
-
-        // Visual state of the box: installed → always on (and locked); a free
-        // name → reflects the pending checkbox; taken → off (and locked).
-        let (checked, enabled) = match &st {
-            shellcmd::State::Installed(_) => (true, false),
-            shellcmd::State::Available(_) => (self.onboarding_install_cmd, true),
-            _ => (false, false),
-        };
-        let label = match &st {
-            shellcmd::State::Installed(n) => {
-                format!("Shell command installed — run `{n}` in any terminal")
-            }
-            shellcmd::State::Available(n) => {
-                format!("Install `{n}` command — open Kyde from any terminal")
-            }
-            _ => "`ky` and `kyde` are already taken on your PATH — skipped".to_string(),
-        };
-
-        let checkbox = div()
-            .size_4()
-            .rounded_sm()
-            .border_1()
-            .border_color(if checked { t.primary } else { t.bg_light })
-            .when(checked, |d| d.bg(t.primary))
-            .flex()
-            .items_center()
-            .justify_center()
-            .when(checked, |d| {
-                d.child(
-                    div()
-                        .text_color(gpui::white())
-                        .text_size(px(11.0))
-                        .child("✓"),
-                )
-            });
-
-        let mut row = div()
-            .flex()
-            .flex_row()
-            .items_center()
-            .gap_2()
-            .pt_3()
-            .mt_1()
-            .border_t_1()
-            .border_color(t.divider)
-            .child(checkbox)
-            .child(
-                div()
-                    .text_color(if enabled {
-                        t.secondary_text
-                    } else {
-                        t.line_number
-                    })
-                    .child(SharedString::from(label)),
-            );
-
-        if enabled {
-            row = row.cursor_pointer().on_mouse_down(
-                MouseButton::Left,
-                cx.listener(|this, _e, _w, cx| {
-                    this.onboarding_install_cmd = !this.onboarding_install_cmd;
-                    cx.notify();
-                }),
-            );
-        }
-
-        let mut col = div().flex().flex_col().gap_1().child(row);
-        if let Some(err) = &self.shell_cmd_error {
-            col = col.child(
-                div()
-                    .text_color(t.status_deleted)
-                    .text_size(px(12.0))
-                    .child(SharedString::from(err.clone())),
-            );
-        }
-        col.into_any_element()
-    }
 
 
 
@@ -2881,237 +2009,11 @@ impl Kyde {
 
 
 
-    /// "Create New Branch" dialog body (own native window). Name field (`branch_query`) +
-    /// Checkout / Overwrite toggles + Cancel / Create.
-    pub(crate) fn render_new_branch_body(&mut self, cx: &mut Context<Self>) -> gpui::AnyElement {
-        let ui = theme::font::UI_FAMILY;
-        let t = theme::get();
-        let (checkout, overwrite) = (self.new_branch_checkout, self.new_branch_overwrite);
-
-        let name_row = div()
-            .flex()
-            .flex_row()
-            .items_center()
-            .gap_3()
-            .px_4()
-            .pt_4()
-            .child(div().flex_none().text_color(t.text).child("Branch Name:"))
-            .child(
-                div()
-                    .flex_1()
-                    .min_w_0()
-                    .px_2()
-                    .py_1p5()
-                    .rounded_md()
-                    .border_1()
-                    .border_color(t.primary)
-                    .bg(t.main_bg)
-                    .font_family(theme::font::FAMILY)
-                    .child(self.branch_query.clone()),
-            );
-
-        let check = |label: &'static str, on: bool, toggle: fn(&mut Self, &mut Context<Self>)| {
-            div()
-                .flex()
-                .flex_row()
-                .items_center()
-                .gap_2()
-                .cursor_pointer()
-                .child(checkbox_box(on))
-                .child(label)
-                .on_mouse_down(
-                    MouseButton::Left,
-                    cx.listener(move |this, _e, _w, cx| toggle(this, cx)),
-                )
-        };
-        let checks = div()
-            .flex()
-            .flex_row()
-            .items_center()
-            .gap_6()
-            .px_4()
-            .pt_3()
-            .child(check("Checkout branch", checkout, |this, cx| {
-                this.new_branch_checkout = !this.new_branch_checkout;
-                cx.notify();
-            }))
-            .child(check("Overwrite existing branch", overwrite, |this, cx| {
-                this.new_branch_overwrite = !this.new_branch_overwrite;
-                cx.notify();
-            }));
-
-        let footer = div()
-            .flex()
-            .flex_row()
-            .justify_end()
-            .gap_2()
-            .px_4()
-            .pb_4()
-            .pt_4()
-            .child(
-                btn_secondary("newbranch-cancel", "Cancel")
-                    .py_1p5()
-                    .on_mouse_down(
-                        MouseButton::Left,
-                        cx.listener(|this, _e, _w, cx| {
-                            this.close_modal_window(ModalKind::NewBranch, cx)
-                        }),
-                    ),
-            )
-            .child(btn_primary("create", "Create").py_1p5().on_mouse_down(
-                MouseButton::Left,
-                cx.listener(|this, _e, _w, cx| this.do_create_branch(cx)),
-            ));
-
-        div()
-            .size_full()
-            .flex()
-            .flex_col()
-            .bg(t.panel_bg)
-            .font_family(ui)
-            .text_size(px(theme::get().ui_font_size))
-            .text_color(t.text)
-            .child(name_row)
-            .child(checks)
-            .child(div().flex_1()) // spacer → footer sits at the bottom
-            .child(footer)
-            .into_any_element()
-    }
-
-
-    /// Body of the "Clear Data & Restart" confirmation modal window. Destructive: wipes the
-    /// whole config dir and restarts.
-    pub(crate) fn render_clear_data_body(&mut self, cx: &mut Context<Self>) -> gpui::AnyElement {
-        let ui = theme::font::UI_FAMILY;
-        let t = theme::get();
-        let cancel = btn_secondary("clear-cancel", "Cancel").on_mouse_down(
-            MouseButton::Left,
-            cx.listener(|this, _e, _w, cx| this.close_modal_window(ModalKind::ClearData, cx)),
-        );
-        let confirm = btn_primary("clear-confirm", "Clear & Restart").on_mouse_down(
-            MouseButton::Left,
-            cx.listener(|this, _e, _w, cx| this.do_clear_data(cx)),
-        );
-        div()
-            .size_full()
-            .flex()
-            .flex_col()
-            .gap_3()
-            .p_4()
-            .font_family(ui)
-            .text_size(px(theme::get().ui_font_size + 1.0))
-            .child(div().text_color(t.text).child("Clear all data & restart?"))
-            .child(
-                div()
-                    .flex_1()
-                    .text_color(t.secondary_text)
-                    .text_size(px(theme::get().ui_font_size))
-                    .child(
-                        "Uninstalls all language plugins and clears cached settings (keymap, \
-                         theme, recent projects, preferences). Kyde will restart. Can't be undone.",
-                    ),
-            )
-            .child(
-                div()
-                    .flex()
-                    .flex_row()
-                    .justify_end()
-                    .gap_2()
-                    .child(cancel)
-                    .child(confirm),
-            )
-            .into_any_element()
-    }
 
 
 
-    /// Font specimen viewer: each bundled family at every available weight, previewing a
-    /// full sentence at a large size so weight differences are obvious.
-    /// Body of the Fonts modal window (hosted by `ModalWindow`, native titlebar).
-    pub(crate) fn render_fonts_body(&mut self, _cx: &mut Context<Self>) -> gpui::AnyElement {
-        let ui = theme::font::UI_FAMILY;
-        let t = theme::get();
-        const QUOTE: &str = "Me fail English? That's unpossible.";
-        let line = |family: &'static str, label: &'static str, weight: FontWeight| {
-            div()
-                .flex()
-                .flex_col()
-                .gap_1()
-                .px_3()
-                .py_2()
-                .child(
-                    div()
-                        .font_family(ui)
-                        .text_color(t.line_number)
-                        .text_size(px(11.0))
-                        .child(SharedString::from(label)),
-                )
-                .child(
-                    div()
-                        .font_family(family)
-                        .font_weight(weight)
-                        .text_size(px(26.0))
-                        .text_color(t.text)
-                        .child(QUOTE),
-                )
-        };
-        let header = |title: &'static str| {
-            div()
-                .px_3()
-                .pt_3()
-                .pb_1()
-                .font_family(ui)
-                .font_weight(FontWeight::SEMIBOLD)
-                .text_color(t.secondary_text)
-                .text_size(px(12.0))
-                .child(SharedString::from(title))
-        };
-        // Fills the modal window (chrome + native "Fonts" titlebar from `ModalWindow`).
-        div()
-            .size_full()
-            .flex()
-            .flex_col()
-            .font_family(ui)
-            .child(
-                div()
-                    .id("fonts-list")
-                    .overflow_y_scroll()
-                    .flex_1()
-                    .flex()
-                    .flex_col()
-                    .pb_2()
-                    .child(header("Inter — UI"))
-                    .child(line(
-                        theme::font::UI_FAMILY,
-                        "Regular · 400",
-                        FontWeight::NORMAL,
-                    ))
-                    .child(line(
-                        theme::font::UI_FAMILY,
-                        "Medium · 500",
-                        FontWeight::MEDIUM,
-                    ))
-                    .child(line(
-                        theme::font::UI_FAMILY,
-                        "SemiBold · 600",
-                        FontWeight::SEMIBOLD,
-                    ))
-                    .child(line(theme::font::UI_FAMILY, "Bold · 700", FontWeight::BOLD))
-                    .child(header("JetBrains Mono — Code"))
-                    .child(line(
-                        theme::font::FAMILY,
-                        "Regular · 400",
-                        FontWeight::NORMAL,
-                    ))
-                    .child(line(
-                        theme::font::FAMILY,
-                        "SemiBold · 600",
-                        FontWeight::SEMIBOLD,
-                    ))
-                    .child(line(theme::font::FAMILY, "Bold · 700", FontWeight::BOLD)),
-            )
-            .into_any_element()
-    }
+
+
 
     /// Bottom terminal panel: a drag-resize divider, a tab strip (one tab per shell +
     /// a new-tab button), and the active terminal filling the rest. Inset to align
